added option to not process json files

+++ track_again.py
@@ -54,12 +54,12 @@ def run(
         max_det=1000,  # maximum detections per image
         device='',  # cuda device, i.e. 0 or 0,1,2,3 or cpu
         show_vid=True,  # show results
-        save_JSON=True, # save results to *.json
-        save_txt=True,  # save results to *.txt
+        save_JSON=False, # save results to *.json
+        save_txt=False,  # save results to *.txt
         save_conf=False,  # save confidences in --save-txt labels
         save_crop=False,  # save cropped prediction boxes
         save_trajectories=False,  # save trajectories for each track
-        save_vid=True,  # save confidences in --save-txt labels
+        save_vid=False,  # save confidences in --save-txt labels
         classes=None,  # filter by class: --class 0, or --class 0 2 3
         agnostic_nms=False,  # class-agnostic NMS
         line_thickness=2,  # bounding box thickness (pixels)
@@ -78,7 +78,6 @@ def run(
     model = AutoBackend(yolo_weights, device=device, dnn=dnn, fp16=half)
     stride, names, pt = model.stride, model.names, model.pt
     imgsz = check_imgsz(imgsz, stride=stride)  # check image size
-    print(imgsz)
     # Dataloader
     bs = 1
     dataset = LoadImages(
@@ -90,12 +89,13 @@ def run(
         vid_stride=vid_stride
     )
     vid_writer = [None]
-    
-    json_file_path = str(outfolder / "track_output.json")
-    with open(json_file_path, 'w') as file:
-        file.write(json.dumps(json_track_out, indent = 4))
-    json_file = open(json_file_path, 'r+')
-    ujson_data = ujson.load(json_file)
+    if save_JSON:
+        
+        json_file_path = str(outfolder / "track_output.json")
+        with open(json_file_path, 'w') as file:
+            file.write(json.dumps(json_track_out, indent = 4))
+        json_file = open(json_file_path, 'r+')
+        ujson_data = ujson.load(json_file)
 
     vid_path = str(outfolder / "track_vid")  # im.jpg, vid.mp4, ...
     vid_path = str(Path(vid_path).with_suffix('.mp4'))  # force *.mp4 suffix on results videos
@@ -243,9 +243,10 @@ def run(
                 
                 # with open(json_file_path, "w") as outfile:
                 #     outfile.write(json.dumps(json_track_out, indent = 4))
-                json_file.seek(0)
-                json_file.truncate()
-                ujson.dump(ujson_data, json_file)
+                if save_JSON:
+                    json_file.seek(0)
+                    json_file.truncate()
+                    ujson.dump(ujson_data, json_file)
 
         else:
             pass
@@ -278,9 +279,9 @@ def run(
         # Print total time (preprocessing + inference + NMS + tracking)
         LOGGER.info(f"{s}{'' if len(detections) else '(no detections), '}{sum([dt.dt for dt in dt if hasattr(dt, 'dt')]) * 1E3:.1f}ms, loop time is: {int((time.time()-start_loop)*1000)}")
         
-    
-    json_file.close()
-    json_file = None
+    if save_JSON:
+        json_file.close()
+        json_file = None
     # Print results
     t = tuple(x.t / seen * 1E3 for x in dt)  # speeds per image
     LOGGER.info(f'Speed: %.1fms pre-process, %.1fms inference, %.1fms NMS, %.1fms {tracking_method} update per image at shape {(1, 3, *imgsz)}' % t)
@@ -296,7 +297,7 @@ def shutdown_function(signal, frame):
         print("Closing JSON file")
         json_file.close()
         json_file = None
-    if(vid_writer is not None):
+    if(vid_writer is not None and hasattr(vid_writer,"release")):
         print("Closing video writer")
         vid_writer.release()
         vid_writer = None
